chore(project): reorganization project structure
